Barebone gemini client
While it works, there are some flaws.
User input is not implemented, neither is support for client
certificates. There is no check for \r\n in the end of the first line
of a response, so it will work with servers that are not implementing
the protocol properly.

I also think that a refactoring is needed if I want to implement user
input.

Anyway, let's see how far I can go with GUI